+++ src/packages/tyhUi/packages/back-top/TyhBackTop.vue
@@ -1,6 +1,6 @@
 <template>
   <div
-    v-show="backShow"
+    v-show="isShow"
     class="tyh-back-top"
     :style="{ bottom, right }"
     @click="toTop"
@@ -11,7 +11,7 @@
 
 <script setup>
 import { ref, onMounted } from 'vue'
-defineProps({
+const props = defineProps({
   bottom: {
     type: String,
     default: '50px'
@@ -19,11 +19,22 @@ defineProps({
   right: {
     type: String,
     default: '50px'
+  },
+  behavior: {
+    type: String,
+    default: 'smooth',
+    validator (v) {
+      return ['smooth', 'instant'].includes(v)
+    }
+  },
+  beyond: {
+    type: Number,
+    default: 100
   }
 })
-const { backShow, toTop } = _TyhBackTop()
+const { isShow, toTop } = _TyhBackTop()
 function _TyhBackTop () {
-  const backShow = ref(false)
+  const isShow = ref(false)
   const handleScroll = () => {
     let timer = null
     return function () {
@@ -31,22 +42,13 @@ function _TyhBackTop () {
         clearTimeout(timer)
       }
       timer = setTimeout(() => {
-        let scrollTop =
-          document.documentElement.scrollTop || document.body.scrollTop
-        scrollTop > 100 ? (backShow.value = true) : (backShow.value = false)
+        let scrollTop = document.documentElement.scrollTop
+        scrollTop > props.beyond ? (isShow.value = true) : (isShow.value = false)
       }, 188)
     }
   }
-  const toTop = () => {
-    let top = document.documentElement.scrollTop || document.body.scrollTop
-    const timeTop = setInterval(() => {
-      document.body.scrollTop = document.documentElement.scrollTop = top -= 50
-      if (top <= 0) {
-        clearInterval(timeTop)
-      }
-    }, 11)
-  }
-  onMounted(() => window.addEventListener('scroll', handleScroll()))
-  return { backShow, toTop }
+  const toTop = () => scrollTo({ top: 0, behavior: props.behavior })
+  onMounted(() => addEventListener('scroll', handleScroll()))
+  return { isShow, toTop }
 }
 </script>